requirements.txt
astroid==2.4.2
autopep8==1.5.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.1
Pillow==7.2.0
pkg-resources==0.0.0
pycodestyle==2.6.0
pylint==2.5.3
RPi.GPIO==0.7.0
six==1.15.0
spidev==3.5
ST7789==0.0.2
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
